\name{shinyepico-package}
\alias{shinyepico-package}
\alias{shinyepico}
\docType{package}
\title{
  shinyÉPICo: a graphical and interactive pipeline to analyze Illumina DNA Methylation Arrays
}
\description{
  ShinyÉPICo is a web interface based on Shiny that makes it easy to do differentially methylated positions (DMP) and differentially methylated regions (DMR) analysis from Illumina EPIC or 450k DNA methylation arrays. This program allows following a standard pipeline of normalization (with minfi package), model creation and statistical analysis (with limma package), and DMR calculation (with mCSEA package) with different options in each step and plots to be able to choose properly. Moreover, you can select different options in the final heatmap and download an RMarkdown report with all the steps chosen.
}
\details{
  The package contains two vignettes that describe throughly its features and options.
}
\keyword{ package }
