% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_shinyepico}
\alias{run_shinyepico}
\title{Run Shiny ÉPICo!}
\usage{
run_shinyepico(
  n_cores = if (parallel::detectCores() > 1) {     cores =
    round(parallel::detectCores()/2, digits = 0) } else {     cores = 1 },
  max_upload_size = 500
)
}
\arguments{
\item{n_cores}{Number of cores to be used in parallelized operations in the application. By default, half of your CPU cores. As more cores mean more RAM used, we don't recommend use more than 1 core per 4GB of RAM available}

\item{max_upload_size}{The limit in MB of the .zip file size to be uploaded. By default, 500MB.}
}
\value{
None
}
\description{
ShinyÉPICo! Interactive minfi and limma pipeline for Illumina methylation arrays
}
\examples{
\dontrun{run_shinyepico()}

}
