% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_shinyepico}
\alias{run_shinyepico}
\title{Run Shiny ÉPICo!}
\usage{
run_shinyepico(
  n_cores = parallel::detectCores()/2,
  max_upload_size = 2000,
  host = "127.0.0.1",
  port = NULL
)
}
\arguments{
\item{n_cores}{Number of cores to be used in parallelized operations in the application. By default, half of your CPU cores. Parallelization affects only to mean and differences calculation and it does not suppose a significant memory overhead.}

\item{max_upload_size}{The limit in MB of the .zip file size to be uploaded. By default, 2000MB.}

\item{host}{IP used to deploy the server. By default, your local IP (127.0.0.1)}

\item{port}{Port used to deploy the server.}
}
\value{
None
}
\description{
ShinyÉPICo! Interactive minfi and limma pipeline for Illumina methylation arrays
}
\examples{
{if(interactive()){run_shinyepico()}}
}
